+++ qabelbox/src/main/java/de/qabel/qabelbox/chat/interactor/MarkAsRead.kt
@@ -1,0 +1,8 @@
+package de.qabel.qabelbox.chat.interactor
+
+
+interface MarkAsRead {
+    fun all(identityKey: String)
+    fun forContact(identityKey: String, contactKey: String)
+}
+