+++ qabelbox/src/main/java/de/qabel/qabelbox/chat/interactor/MessageStateBroadcaster.kt
@@ -1,0 +1,8 @@
+package de.qabel.qabelbox.chat.interactor
+
+import de.qabel.core.config.Identity
+
+interface MessageStateBroadcaster {
+    fun messagesRead(identity: Identity)
+}
+