+++ qabelbox/src/main/java/de/qabel/qabelbox/chat/services/AndroidChatService.kt
@@ -6,6 +6,7 @@ import de.qabel.qabelbox.QabelBoxApplication
 import de.qabel.qabelbox.QblBroadcastConstants
 import de.qabel.qabelbox.QblBroadcastConstants.Chat.Service
 import de.qabel.qabelbox.chat.interactor.ChatServiceUseCase
+import de.qabel.qabelbox.chat.interactor.MarkAsRead
 import de.qabel.qabelbox.chat.notifications.ChatNotificationManager
 import de.qabel.qabelbox.chat.transformers.ChatMessageTransformer
 import de.qabel.qabelbox.helper.Helper
@@ -28,6 +29,7 @@ open class AndroidChatService() : IntentService(AndroidChatService::class.java.s
     private fun Intent.identityKey(): String = getStringExtra(PARAM_IDENTITY_KEY)!!
 
     @Inject lateinit var chatService: ChatServiceUseCase
+    @Inject lateinit var markAsRead: MarkAsRead
     @Inject lateinit var chatMessageTransformer: ChatMessageTransformer
     @Inject lateinit var chatNotificationManager: ChatNotificationManager
     @Inject lateinit var crashReporter: CrashReporter
@@ -66,7 +68,7 @@ open class AndroidChatService() : IntentService(AndroidChatService::class.java.s
         val identityKey = intent.identityKey()
         val contactKey = intent.contactKey()
         chatService.ignoreContact(identityKey, contactKey)
-        chatService.markContactMessagesRead(identityKey, contactKey)
+        markAsRead.forContact(identityKey, contactKey)
         chatNotificationManager.hideNotification(identityKey, contactKey)
         sendChatStateChanged()
         sendContactsUpdated()
@@ -76,10 +78,10 @@ open class AndroidChatService() : IntentService(AndroidChatService::class.java.s
         val identityKey = intent.identityKey()
         if (intent.hasExtra(PARAM_CONTACT_KEY)) {
             val contactKey = intent.contactKey()
-            chatService.markContactMessagesRead(identityKey, contactKey)
+            markAsRead.forContact(identityKey, contactKey)
             chatNotificationManager.hideNotification(identityKey, contactKey)
         } else {
-            chatService.markIdentityMessagesRead(identityKey)
+            markAsRead.all(identityKey)
             chatNotificationManager.hideNotification(identityKey, null)
         }
         sendChatStateChanged()