+++ qabelbox/src/main/java/de/qabel/qabelbox/index/dagger/IndexSearchModule.java
@@ -2,14 +2,6 @@ package de.qabel.qabelbox.index.dagger;
 
 import dagger.Module;
 import dagger.Provides;
-import de.qabel.core.repository.ContactRepository;
-import de.qabel.core.repository.IdentityRepository;
-import de.qabel.qabelbox.chat.interactor.ChatOverviewUseCase;
-import de.qabel.qabelbox.chat.interactor.MainChatOverviewUseCase;
-import de.qabel.qabelbox.chat.transformers.ChatMessageTransformer;
-import de.qabel.qabelbox.chat.view.presenters.ChatOverviewPresenter;
-import de.qabel.qabelbox.chat.view.presenters.MainChatOverviewPresenter;
-import de.qabel.qabelbox.chat.view.views.ChatOverview;
 import de.qabel.qabelbox.dagger.scopes.ActivityScope;
 import de.qabel.qabelbox.index.interactor.IndexSearchUseCase;
 import de.qabel.qabelbox.index.interactor.MainIndexSearchUseCase;